feat: add useRouteAssetId hook
